requirements.txt
# This file is here because many Platforms as a Service look for
#	requirements.txt in the root directory of a project.
pylibmc==1.4.2
djangorestframework=1.5.6
django-heroku-memcacheify==0.8

-r requirements/production.txt
